+++ NEWS.md
@@ -1,2 +1,6 @@
 # Changes in version 0.99.0 (2022-03-28)
 - Submitted to Bioconductor
+
+# Changes in version 0.99.1 (2022-04-01)
+- Updated vignette to use pre-computed output
+- Allow removal of isolated nodes in `plot_gsNetwork`